// as201-ms-person-fe-develop/src/app/models/studentModel.ts
export class studentModel{
    id?: number | undefined;
    person_id?: number | undefined;
    person_name?: string | undefined;
    career_id?: number | undefined;
    career_name?: string | undefined;
    institutional_email?: string | undefined;
    pay_method?: string | undefined;
    admission_date?: string | undefined;
    guardian_name?: string | undefined;
    home_phone?: number | undefined;
    status?: string | undefined;
    constructor() {}
}